reference errors

+++ GPRO-Graphics1/include/gpro/hittable_list.h
@@ -29,16 +29,16 @@ public:
     void add(shared_ptr<hittable> object) { objects.push_back(object); }
 
     virtual bool hit(
-        const ray& r, double tmin, double tmax, hit_record& rec) const override;
+        const ray& r, float tmin, float tmax, hit_record& rec) const override;
 
 public:
     std::vector<shared_ptr<hittable>> objects;
 };
 
-bool hittable_list::hit(const ray& r, double t_min, double t_max, hit_record& rec) const {
+bool hittable_list::hit(const ray& r, float t_min, float t_max, hit_record& rec) const {
     hit_record temp_rec;
     bool hit_anything = false;
-    double closest_so_far = t_max;
+    float closest_so_far = t_max;
 
     for (int i = 0; i < t_max; i++) {
         if (hit(r, t_min, closest_so_far, temp_rec)) {
